ops: align surface policy and reduce file-size gate backlog
