+++ docs/workspace/codex_enforcer.md
@@ -89,17 +89,19 @@ Completion requires all of the following:
 - If explicit permission is missing for any file-type change or language migration, mark the task `BLOCKED — missing explicit file-type migration permission` and stop.
 
 ## Repository Placement Rules (Mandatory)
-- Canonical code locations are limited to: `core/`, `client/`, `tests/`, and `adapters/`.
+- Canonical code locations are limited to: `core/`, `surface/`, `client/`, `tests/`, and `adapters/`.
 - `apps/` is app-only. It may contain only standalone apps that run on top of the client/runtime boundary.
-- Any path under `apps/` must be deletable without changing core/client/adapters/tests behavior.
-- System code must not import from `apps/**`. If system code needs shared logic, move that logic into `core/`, `client/`, `tests/`, or `adapters/` first.
+- Any path under `apps/` must be deletable without changing core/surface/client/adapters/tests behavior.
+- System code must not import from `apps/**`. If system code needs shared logic, move that logic into `core/`, `surface/`, `client/`, `tests/`, or `adapters/` first.
 - `apps/` is never a script/tool dump. Shared helpers, wrappers, and runtime bridges are prohibited in `apps/`.
 - Top-level `scripts/` is prohibited. Do not create or reintroduce it.
 - CI/dev/test tooling scripts must live under `tests/tooling/scripts/`.
 - Runtime/operator utilities must live under `client/runtime/systems/**` (or `core/**` when authoritative).
 - If initialization/bootstrap installers need a dedicated surface, use `setup/` as the only root-level exception.
 - Placement decision rule:
-  - system authority/runtime path => `core/` (or `client/runtime/systems/**` only as thin runtime/client surface)
+  - system authority/runtime path => `core/`
+  - orchestration coordination path (non-authoritative) => `surface/orchestration/**`
+  - client runtime wrappers/UX path => `client/runtime/systems/**` (thin runtime/client surface only)
   - developer/user operational scripts => `client/`
   - test/CI tooling => `tests/`
   - integration bridges for external software => `adapters/`
